requirements.txt
pdfminer==20140328
peewee>=2.5
Scrapy>=0.24
service-identity>=14.0
gensim>=0.11
nltk==3.0.1
scipy>=0.15
numpy>=1.9
scikit-learn>=0.16
lda>=1.0
beautifulsoup4>=4.3
python-pptx>=0.5
langdetect>=1.0.5
pyvabamorf>=1.6
pathos>=0.2.0